model/requirements.txt
ansi2html==1.8.0
asgiref==3.7.2
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
backports.zoneinfo==0.2.1
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
comm==0.1.3
ConfigArgParse==1.5.5
dash==2.11.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.6.7
decorator==5.1.1
Django==4.2.2
django-cors-headers==4.1.0
djangorestframework==3.14.0
executing==1.2.0
fastjsonschema==2.17.1
filelock==3.9.0
Flask==2.2.5
gunicorn==20.1.0
idna==3.4
imageio==2.31.1
importlib-metadata==6.8.0
importlib-resources==6.0.0
ipykernel==6.24.0
ipython==8.12.2
ipywidgets==8.0.7
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyterlab-widgets==3.0.8
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mpmath==1.2.1
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==3.0
numpy==1.24.4
open3d==0.17.0
packaging==23.1
parso==0.8.3
pickleshare==0.7.5
Pillow==10.0.0
pkgutil_resolve_name==1.3.10
platformdirs==3.9.1
plotly==5.15.0
prompt-toolkit==3.0.39
psutil==5.9.5
pure-eval==0.2.2
Pygments==2.15.1
python-dateutil==2.8.2
pytz==2023.3
pywin32==306
pyzmq==25.1.0
referencing==0.30.0
requests==2.31.0
retrying==1.3.4
rpds-py==0.9.2
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
sympy==1.11.1
tenacity==8.2.2
torch==2.0.1+cu118
torchaudio==2.0.2+cu118
torchvision==0.15.2+cu118
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.3
wcwidth==0.2.6
Werkzeug==2.2.3
widgetsnbextension==4.0.8
zipp==3.16.2
